requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
feedparser==6.0.10
python-dateutil==2.8.2
APScheduler==3.10.4
gunicorn==21.2.0


psycopg2-binary==2.9.7

